setup explicit layouts in API
Transpose based layout handling approach needs to be rewritten for
scenario of explicit layouts. Currently, if set explicit layout for
a tensor connected to node that are all not have implicit layout,
transpose based layout handling approach will skip it because it's
OP centralized. It should be tensor centralized, which wrap the tensor
with transpose, and replace original tensors regarding connected OP.